Module robodrummer::commands

This module defines the command line interface for the application using clap.

Re-exports

pub use crate::tui::start_tui as tui;

Structs

Arguments

ArpeggioArgs

CCArgs

CombinerArgs

CompletionsArgs

DevArgs

DrumArgs

Arguments for the drum output mode

EucledeanArgs

GenerateDataArgs

HyperArgs

MetronomeArgs

MidiBrokerArgs

NPDAGArgs

PolyEuclideanArgs

RobotArgs

RunArgs

TrainArgs

TuiArgs

Enums

BrokerMode

Command

DevCommand

DrumOutput

OutputMode

RhythmAlgorithm

RobotCommand

TrainMode

Constants

FEEL_PORT

METRONOME_PORT

MIDI_PORT

OSC_PORT

OUTPUT_PORT

Functions

broke

handle midi incomping messages

combine

dev

gendata

Generate input-output data to train the reservoir, based on the given arguments

metronome

run

run the selected model, determined by the parameters in args.

train

update_completions