Function robodrummer::data::load_train_data

pub fn load_train_data(
    name: &str,
    timestep: f64,
    input_width: usize,
    target_width: usize,
    shift: Option<usize>
) -> Result<Data, Box<dyn Error>>

Load the training data from a .bin file

The training data is stored in a time-domain format, and needs to be converted to a discrete timestep format. This is where the parameters timestep and input_width come into play.

Arguments

name: The name of the training data file

timestep: The time between each timestep

input_width: The number of timesteps to consider as input

shift: the amount of timesteps to shift the target data into the future