Struct robodrummer::metronomer::inputwindow::InputWindow

pub struct InputWindow {
    pub window: VecDeque<u128>,
    pub best_frequency: f64,
    pub min_band: f64,
    pub max_band: f64,
    pub hit_count: u32,
    /* private fields */
}

The InputWindow keeps memory of the input hits and fourier related data

Fields
window: VecDeque<u128>

The window keeps track of times of hits

example:

first hit at time 100

second hit at time 120

third hit at time 150 then we have the resulting window: (50, 20, 0)

fourth hit at time 170 then we have the resulting window: (70, 50, 20)

best_frequency: f64min_band: f64max_band: f64hit_count: u32

The number of hits

Implementations

impl InputWindow

pub fn new_with_size(size: usize, sample_period: f64) -> Self

pub fn hit(&mut self, action: HitAction) -> bool

Add a hit to the input window, using the given HitAction Returns true if the best frequency was updated

pub fn show_window(&self)

pub fn set_band(&mut self)

pub fn create_fft_buffer(&self) -> Option<Vec<Complex<f64>>>

pub fn plot_decision(&mut self)